DAOS-6222 (PR-5913) proactively add support for dfs mode
Adds support for --mode flag in anticipation of PR-5913 landing.
Should merge cleanly with that PR.

Also pulls flags out into a new file to reduce the size of util.go
