+++ src/control/cmd/daos/util.go
@@ -11,11 +11,8 @@ import (
 	"fmt"
 	"io"
 	"os"
-	"strconv"
-	"strings"
 	"unsafe"
 
-	"github.com/dustin/go-humanize"
 	"github.com/google/uuid"
 	"github.com/pkg/errors"
 
@@ -30,7 +27,8 @@ import (
 #cgo CFLAGS: -I${SRCDIR}/../../../utils
 #cgo LDFLAGS: -lgurt -lcart -ldaos -ldaos_common -lduns -ldfs -luuid -ldaos_cmd_hdlrs
 
-#include <sys/stat.h>
+#define D_LOGFAC	DD_FAC(client)
+
 #include <daos.h>
 #include <daos/common.h>
 
@@ -272,145 +270,6 @@ func resolveDunsPath(path string, ap *C.struct_cmd_args_s) error {
 	if err := daosError(rc); err != nil {
 		return errors.Wrapf(err, "failed to resolve path %s", path)
 	}
-
-	return nil
-}
-
-type labelOrUUID struct {
-	UUID  uuid.UUID
-	Label string
-}
-
-func (f labelOrUUID) Empty() bool {
-	return !f.HasLabel() && !f.HasUUID()
-}
-
-func (f labelOrUUID) HasLabel() bool {
-	return f.Label != ""
-}
-
-func (f labelOrUUID) HasUUID() bool {
-	return f.UUID != uuid.Nil
-}
-
-func (f labelOrUUID) String() string {
-	switch {
-	case f.HasLabel():
-		return f.Label
-	case f.HasUUID():
-		return f.UUID.String()
-	default:
-		return "<no label or uuid set>"
-	}
-}
-
-func (f *labelOrUUID) UnmarshalFlag(fv string) error {
-	uuid, err := uuid.Parse(fv)
-	if err == nil {
-		f.UUID = uuid
-		return nil
-	}
-
-	f.Label = fv
-	return nil
-}
-
-type epochRange struct {
-	set   bool
-	begin uint64
-	end   uint64
-}
-
-func (er *epochRange) String() string {
-	return fmt.Sprintf("%d-%d", er.begin, er.end)
-}
-
-func (er *epochRange) UnmarshalFlag(fv string) error {
-	er.set = true
-	n, err := fmt.Sscanf(fv, "%d-%d", &er.begin, &er.end)
-	if err != nil {
-		return err
-	}
-	if n != 2 {
-		return errors.Errorf("range=%q must be in A-B form", fv)
-	}
-	if er.begin >= er.end {
-		return errors.Errorf("range begin must be < end")
-	}
-	return nil
-}
-
-type chunkSize struct {
-	set  bool
-	size C.uint64_t
-}
-
-func (c *chunkSize) UnmarshalFlag(fv string) error {
-	if fv == "" {
-		return errors.New("empty chunk size")
-	}
-
-	size, err := humanize.ParseBytes(fv)
-	if err != nil {
-		return err
-	}
-	c.size = C.uint64_t(size)
-	c.set = true
-
-	return nil
-}
-
-type objectClass struct {
-	set   bool
-	class C.ushort
-}
-
-func (oc *objectClass) UnmarshalFlag(fv string) error {
-	if fv == "" {
-		return errors.New("empty object class")
-	}
-
-	cObjClass := C.CString(fv)
-	defer freeString(cObjClass)
-
-	oc.class = (C.ushort)(C.daos_oclass_name2id(cObjClass))
-	if oc.class == C.OC_UNKNOWN {
-		return errors.Errorf("unknown object class %q",
-			fv)
-	}
-	oc.set = true
-
-	return nil
-}
-
-type oidFlag struct {
-	set bool
-	oid C.daos_obj_id_t
-}
-
-func (f *oidFlag) String() string {
-	return fmt.Sprintf("%d.%d", f.oid.hi, f.oid.lo)
-}
-
-func (f *oidFlag) UnmarshalFlag(fv string) error {
-	parseErr := fmt.Sprintf("failed to parse %q as OID (expected HI.LO)", fv)
-	comps := strings.Split(fv, ".")
-	if len(comps) != 2 {
-		return errors.New(parseErr)
-	}
-
-	val, err := strconv.ParseUint(comps[0], 10, 64)
-	if err != nil {
-		return errors.Wrap(err, parseErr)
-	}
-	f.oid.hi = C.uint64_t(val)
-	val, err = strconv.ParseUint(comps[1], 10, 64)
-	if err != nil {
-		return errors.Wrap(err, parseErr)
-	}
-	f.oid.lo = C.uint64_t(val)
-
-	f.set = true
 
 	return nil
 }
