+++ src/utils/daos_hdlr.h
@@ -89,6 +89,7 @@ struct cmd_args_s {
 	char			*dst;		/* --dst path for fs copy */
 	daos_cont_layout_t	type;		/* --type cont type */
 	daos_oclass_id_t	oclass;		/* --oclass object class */
+	uint32_t		mode;		/* --posix consistency mode */
 	daos_size_t		chunk_size;	/* --chunk_size of cont objs */
 
 	/* Container snapshot/rollback related */
